Added Game structure
