+++ main.hpp
@@ -11,21 +11,22 @@
 #include <stdlib.h>
 #include <unistd.h>
 
-// Created Libraries
-#include "Board.hpp"
-
 using namespace std;
 
 typedef struct {
   string board_config;
   string game_played;
   string language_file;
+  int players;
+  int this_player_go;
 } options;
+
+// Created Libraries
+#include "Game.hpp"
 
 void print_error(string err);
 void usage(void);
 bool check_options(options o);
 bool check_file_exists(string s, bool print_err = false);
-rapidjson::Document get_config_from_file(string& config);
 
 #endif
